Sushant Sharma's Deliveroo Tech Task Submission
This is a Technical Task given by Deliveroo to Sushant Sharma, to be completed as part of the ongoing interview process. This is about writing a program/application, that takes cron strings as input, and prints the details of the cron command - the minutes, hours, days of month, months and the days of the week that the command is supposed to be executed.